Do not break after first iteration, also check if path is valid

+++ scripts/reset_commits.sh
@@ -1,17 +1,19 @@
 # ./reset_commits.sh [spec file] [repos dir]
 
-while IFS= read -r line
-do
-  split=( $line )
+while IFS= read -r line; do
+  split=($line)
   url=${split[0]}
   commit=${split[1]}
 
   author=$(basename $(dirname $url))
   repo=$(basename $url .git)
 
-  cd $2/$author/$repo
-  git reset --hard $commit
-  cd -
-
-  break
-done < "$1"
+  if [[ -d "$2/$author/$repo" ]]; then
+    echo "$2/$author/$repo" 
+    cd "$2/$author/$repo"
+    git reset --hard $commit
+    cd - > /dev/null
+  else
+    echo "Skipping $2/$author/$repo; folder does not exist"
+  fi
+done <"$1"
